requirements.txt
django-admin
pytorch
transformers
scikit-learn
sentence-transformers
